oauth2/requirements.txt
discord.py
requests_oauthlib
flask
gunicorn
configargparse